+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [0.10.0] (pre-release)
+- Improve progress display during review.
+- Show partial results when cancelling a larger review.
 - The "Enable Debug Output" setting now logs to a separate channel in the output view.
 
 ## [0.9.0]